cmd: export support auto download mod, support pkg/...
