+++ cmd/internal/export/export.go
@@ -19,11 +19,16 @@ package export
 
 import (
 	"bytes"
+	"encoding/json"
 	"fmt"
 	"go/format"
+	"go/types"
 	"io/ioutil"
+	"log"
 	"os"
+	"os/exec"
 	"path/filepath"
+	"strings"
 
 	"github.com/goplus/gop/cmd/internal/base"
 	"github.com/goplus/gop/cmd/internal/export/gopkg"
@@ -50,13 +55,25 @@ func init() {
 	Cmd.Run = runCmd
 }
 
+var (
+	libDir string
+	gobin  string
+)
+
+func init() {
+	var err error
+	gobin, err = exec.LookPath("go")
+	if err != nil {
+		log.Fatalln("not found go bin in PATH", err)
+	}
+}
+
 func runCmd(cmd *base.Command, args []string) {
 	flag.Parse(args)
 	if flag.NArg() < 1 {
 		cmd.Usage(os.Stderr)
 		return
 	}
-	var libDir string
 	if flagExportDir != "" {
 		libDir = flagExportDir
 	} else {
@@ -69,19 +86,144 @@ func runCmd(cmd *base.Command, args []string) {
 	}
 
 	for _, pkgPath := range flag.Args() {
-		err := exportPkg(pkgPath, libDir)
+		var exporAll bool
+		if strings.HasSuffix(pkgPath, "/...") {
+			pkgPath = pkgPath[:len(pkgPath)-4]
+			exporAll = true
+		}
+		pkg, err := LookupPkg(pkgPath)
 		if err != nil {
-			fmt.Fprintf(os.Stderr, "export pkg %q failed, %v\n", pkgPath, err)
+			fmt.Printf("lookup %q source error: %v\n", pkgPath, err)
+			continue
+		}
+		fmt.Printf("source %v\n", pkg.Dir)
+		if exporAll {
+			exportDir(pkg.Dir)
+			continue
+		}
+		if pkg.Goroot {
+			err = exportPkg(pkg.ImportPath, "")
 		} else {
-			fmt.Fprintf(os.Stdout, "export pkg %q success\n", pkgPath)
+			err = exportPkg(pkg.ImportPath, pkg.Dir)
+		}
+		if err != nil {
+			fmt.Printf("export %q error: %v\n", pkg.ImportPath, err)
+		} else {
+			fmt.Printf("export %q success\n", pkg.ImportPath)
 		}
 	}
 }
 
-func exportPkg(pkgPath string, libDir string) error {
-	pkg, err := gopkg.Import(pkgPath)
+type Package struct {
+	Dir        string // directory containing package sources
+	ImportPath string // import path of package in dir
+	Goroot     bool   // is this package in the Go root?
+}
+
+func LookupPkg(pkgPath string) (*Package, error) {
+	// check go list
+	if !strings.Contains(pkgPath, "@") {
+		pkg, err := checkGoList(pkgPath)
+		if err == nil {
+			return pkg, nil
+		}
+	}
+	// check go mod
+	srcDir, err := gopkg.LookupMod(pkgPath)
+	if err == nil {
+		_, pkgPath = gopkg.ParsePkgVer(pkgPath)
+		return &Package{ImportPath: pkgPath, Dir: srcDir}, nil
+	}
+	// check download mod
+	mod, pkg := gopkg.ParsePkgVer(pkgPath)
+	info, err := downloadMod(mod)
 	if err != nil {
+		return nil, fmt.Errorf("download %q failed, %v", mod, err)
+	}
+	return &Package{ImportPath: pkg, Dir: info.Dir}, nil
+}
+
+type ModInfo struct {
+	Path     string // module path
+	Version  string // module version
+	Error    string // error loading module
+	Info     string // absolute path to cached .info file
+	GoMod    string // absolute path to cached .mod file
+	Zip      string // absolute path to cached .zip file
+	Dir      string // absolute path to cached source root directory
+	Sum      string // checksum for path, version (as in go.sum)
+	GoModSum string // checksum for go.mod (as in go.sum)
+}
+
+func downloadMod(pkgPath string) (*ModInfo, error) {
+	fmt.Println("exec: go download mod", pkgPath)
+	cmd := exec.Command(gobin, "mod", "download", "-json", pkgPath)
+	data, err := cmd.Output()
+	if err != nil {
+		return nil, err
+	}
+	var info ModInfo
+	err = json.Unmarshal(data, &info)
+	if err != nil {
+		return nil, err
+	}
+	return &info, nil
+}
+
+func exportDir(dir string) error {
+	cmd := exec.Command(gobin, "list", "-f", "{{.ImportPath}} {{.Dir}}", "./...")
+	cmd.Dir = dir
+	data, err := cmd.CombinedOutput()
+	if err != nil {
+		log.Println(string(data))
 		return err
+	}
+	lines := strings.Split(string(data), "\n")
+	for _, line := range lines {
+		ar := strings.Split(line, " ")
+		if len(ar) == 2 {
+			pkg, dir := ar[0], ar[1]
+			err := exportPkg(pkg, dir)
+			if strings.Contains(pkg, "internal") || strings.Contains(pkg, "vendor") {
+				continue
+			}
+			if err == nil {
+				fmt.Printf("export %q success\n", pkg)
+			} else if err != gopkg.ErrIgnore {
+				fmt.Printf("export %v error,%v\n", pkg, err)
+			}
+		}
+	}
+	return nil
+}
+
+func checkGoList(pkgPath string) (*Package, error) {
+	cmd := exec.Command(gobin, "list", "-json", pkgPath)
+	data, err := cmd.Output()
+	if err != nil {
+		return nil, err
+	}
+	var pkg Package
+	err = json.Unmarshal(data, &pkg)
+	if err != nil {
+		return nil, err
+	}
+	return &pkg, nil
+}
+
+func exportPkg(pkgPath string, srcDir string) (err error) {
+	var pkg *types.Package
+	if srcDir == "" {
+		pkg, err = gopkg.Import(pkgPath)
+	} else {
+		pkg, err = gopkg.ImportSource(pkgPath, srcDir)
+	}
+	if err != nil {
+		log.Printf("import %q failed: %v", pkgPath, err)
+		return err
+	}
+	if pkg.Name() == "main" {
+		return gopkg.ErrIgnore
 	}
 	var buf bytes.Buffer
 	err = gopkg.ExportPackage(pkg, &buf)
@@ -96,7 +238,7 @@ func exportPkg(pkgPath string, libDir string) error {
 	dir := filepath.Join(libDir, pkg.Path())
 	os.MkdirAll(dir, 0777)
 	err = ioutil.WriteFile(filepath.Join(dir, "gomod_export.go"), data, 0666)
-	return err
+	return
 }
 
 // -----------------------------------------------------------------------------
