+++ go.mod
@@ -5,6 +5,4 @@ go 1.14
 require (
 	github.com/peterh/liner v1.2.0
 	github.com/qiniu/x v1.11.5
-	github.com/visualfc/fastmod v1.3.6 // indirect
-	github.com/visualfc/gotools v1.3.3 // indirect
 )
